ap/requirements.txt
Django==1.6.5
Django-Select2==4.2.2
Pillow==2.5.1
PyExecJS==1.0.4
PyPDF2==1.22
PyReact==0.3.1
Pygments==1.6
South==1.0
coverage==3.6
dj-database-url==0.3.0
dj-static==0.0.6
django-adminactions==0.5
django-autocomplete-light==1.4.14
django-autofixture==0.8.0
django-bootstrap3==4.2.0
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.4.0
django-debug-toolbar==1.2.1
django-discover-runner==1.0
django-extensions==1.3.8
django-grappelli==2.5.3
django-nose==1.2
django-report-builder==2.0.1
django-report-utils==0.2.5
django-reset==0.2.0
django-tables2==0.15.0
django-tastypie==0.11.1
django-toolbelt==0.0.1
djangorestframework==2.3.14
djangorestframework-bulk==0.1.3
djorm-ext-core==0.7
djorm-ext-expressions==0.6
djorm-pgarray==1.0.2
gunicorn==19.0.0
html5lib==0.999
jdcal==1.0
nose==1.3.3
openpyxl==2.0.4
pdfminer==20140328
psycopg2==2.5.3
pudb==2014.1
pyPdf==1.13
python-dateutil==2.2
python-mimeparse==0.1.4
python-placeholder==0.2
pytz==2014.4
reportlab==3.1.8
six==1.7.3
sqlparse==0.1.11
static3==0.5.1
unicodecsv==0.9.4
urwid==1.2.1
wsgiref==0.1.2
xhtml2pdf==0.0.6
xlrd==0.9.3
xlwt==0.7.5
